+++ inject/inject-modules/src/test/scala/com/twitter/inject/modules/InMemoryStatsReceiverModule.scala
@@ -8,4 +8,7 @@ object InMemoryStatsReceiverModule extends TwitterModule {
   override def configure(): Unit = {
     bind[StatsReceiver].to[InMemoryStatsReceiver].in[Singleton]
   }
+
+  /**  Java-friendly way to access this module as a singleton instance */
+  def get(): this.type = this
 }