+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Unreleased
 Added
 ~~~~~
 
+* inject-app: Add more Java-friendly constructors for the TestInjector. ``PHAB_ID=D520900``
+
 Changed
 ~~~~~~~